Converting Activity consumption parameter to Long

See merge request cse1105/2021-2022/team-repositories/oopp-group-51/repository-template!71

+++ client/src/main/java/client/scenes/GameCtrl.java
@@ -241,16 +241,16 @@ public abstract class GameCtrl implements Initializable {
         }
     }
 
-    private void renderEstimationAnswers(List<Integer> correctAnswers) {
-        int givenAnswer = 0;
-        int actualAnswer = correctAnswers.get(0);
+    private void renderEstimationAnswers(List<Long> correctAnswers) {
+        long givenAnswer = 0L;
+        long actualAnswer = correctAnswers.get(0);
         try {
-            givenAnswer = Integer.parseInt(estimationAnswer.getText());
+            givenAnswer = Long.parseLong(estimationAnswer.getText());
         } catch (NumberFormatException ex) {
             givenAnswer = actualAnswer;
         }
 
-        int diff = givenAnswer - actualAnswer;
+        long diff = givenAnswer - actualAnswer;
         String correctAnswer = "Correct Answer: " + actualAnswer;
 
         if (diff > 0) {
@@ -269,9 +269,9 @@ public abstract class GameCtrl implements Initializable {
      *
      * @param correctIndices Indexes of the correct answer(s)
      */
-    protected void renderMultipleChoiceAnswers(List<Integer> correctIndices) {
+    protected void renderMultipleChoiceAnswers(List<Long> correctIndices) {
         for (int i = 0; i < multiChoiceAnswers.size(); ++i) {
-            if (correctIndices.contains(i)) {
+            if (correctIndices.contains((long) i)) {
                 multiChoiceAnswers.get(i).setStyle("-fx-background-color: green");
             } else {
                 multiChoiceAnswers.get(i).setDisable(true);
@@ -452,7 +452,7 @@ public abstract class GameCtrl implements Initializable {
                 break;
             case RANGE_GUESS:
                 try {
-                    ans.addAnswer(Integer.parseInt(estimationAnswer.getText()));
+                    ans.addAnswer(Long.parseLong(estimationAnswer.getText()));
                 } catch (NumberFormatException ex) {
                     System.out.println("Invalid answer yo");
                     if (!initiatedByTimer) {
